+++ src/Addons/34thPRC_ArmourCustom/config.cpp
@@ -90,6 +90,7 @@ class CfgWeapons
 
 	class 34thPRC_ArmourStandard_EOD_Helmet_Urban;
 	class 34thPRC_ArmourStandard_JFO_Helmet_Urban;
+	class 34thPRC_ArmourStandard_Recon_Helmet_Urban;
 
 	//includes
 	#include "data\config_weapons.hpp"